[refactor] enum 등 상수 클래스 폴더 생성 및 이동
