+++ BE/src/main/java/cuketmon/constant/type/Type.java
@@ -1,4 +1,4 @@
-package cuketmon.type;
+package cuketmon.constant.type;
 
 import java.util.Arrays;
 import lombok.Getter;
