+++ BE/src/main/java/cuketmon/constant/type/TypeDamageTable.java
@@ -1,4 +1,4 @@
-package cuketmon.type;
+package cuketmon.constant.type;
 
 import java.util.ArrayList;
 import java.util.List;
