+++ BE/src/main/java/cuketmon/skill/prompt/repository/PromptRepository.java
@@ -1,6 +1,6 @@
-package cuketmon.prompt.repository;
+package cuketmon.skill.prompt.repository;
 
-import cuketmon.prompt.entity.Prompt;
+import cuketmon.skill.prompt.entity.Prompt;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
